- FEAT: finish server arguments parsing
- FEAT: link parsing options functions with the server
- ADD: add a first version of option functions in the server
- FEAT: big improvement on the lib with type changes on functions and data structures to be cleaner
- ADD: start to create some parsing functions
- ADD: start a first architecture for the server files
- FEAT: improve the libmy C (yes again)
- FIX: remove some work in the create str to remove compilation errors
- FEAT: add new logs macros
- STYLE: fix coding-style errors with libmy C
- FEAT: implementing logs features in the libmy C
- STYLE: fix coding style errors with documentation
- FEAT: start a first server architecture by adding new data structures
- FEAT: update and improve the libmy C

+++ libs/my_c/src/array/my_array_len.c
@@ -15,12 +15,12 @@
 /**
  * @brief Returns the length of an array (array)
  * @param array The array to check
- * @return <b>int</b> The length of the array
+ * @return <b>size_t</b> The length of the array
  * @author Nicolas TORO
  */
-int my_array_len(void **array)
+size_t my_array_len(void **array)
 {
-    int len = 0;
+    size_t len = 0;
 
     if (array == NULL)
         return 0;
