+++ libs/my_c/src/array/my_word_array_len.c
@@ -16,16 +16,16 @@
  * @brief Returns the length of an word array (array)
  * @param array The array to check
  * @param nb_str The number of strings in the array
- * @return <b>int</b> The length of the word array
+ * @return <b>size_t</b> The length of the word array
  * @author Nicolas TORO
  */
-int my_word_array_len(char **array, int nb_str)
+size_t my_word_array_len(char **array, size_t nb_str)
 {
-    int len = 0;
+    size_t len = 0;
 
     if (array == NULL)
         return 0;
-    for (int i = 0; i < nb_str; i++)
+    for (size_t i = 0; i < nb_str; i++)
         len += my_strlen(array[i]);
     return len;
 }
